+++ CHANGES.md
@@ -4,11 +4,11 @@
 
 Features
 ~~~~~~~~
+- pat-validation: Allow for HTML5 style ``required`` attributes without a value.
 - pat-validation: Added the possibility to check for fields equality
 - pat-validation: Dont use :input jquery extension for better performance
 - pat-validation: Update validate.js to 0.13.1
 - Prevent "Modernizr.inputtypes is undefined" error
-
 - pat-validation: Do not trigger on empty dates. Fixes: #711
 - Remove PhantomJS - we're using ChromeHeadless already.
 - Simplify package.json and remove unused.